requirements_cron.txt
# Cron Job 專用依賴 - 簡化版本
pandas>=1.5.0
openpyxl>=3.1.0
google-cloud-storage>=2.10.0
